static and order methods added

+++ src/app/modules/user/user.service.ts
@@ -7,6 +7,9 @@ const getUsers = async () => {
 };
 
 const getSingleUser = async (id: string) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
   const result = await User.findOne({ userId: Number(id) });
   return result;
 };
@@ -18,16 +21,44 @@ const createUser = async (user: TUser) => {
 };
 
 const editUser = async (id: string, data: object) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
+  // const doc = await User.find({ userId: Number(id) });
+  // if (data?.hobbies) {
+  //   doc.hobbies = [...doc.hobbies, ...data.hobbies];
+  //   console.log(doc.hobbies);
+  // }
   const result = await User.findOneAndUpdate({ userId: Number(id) }, data, {
     new: true,
   });
-  console.log(result);
   return result;
 };
 
 const deleteUser = async (id: string) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
   const result = await User.deleteOne({ userId: Number(id) });
   return result;
+};
+
+const addOrder = async (id: string) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
+};
+
+const allOrders = async (id: string) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
+};
+
+const totalPrice = async (id: string) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("user does not exist");
+  }
 };
 
 export const userServices = {
@@ -36,4 +67,7 @@ export const userServices = {
   createUser,
   editUser,
   deleteUser,
+  addOrder,
+  allOrders,
+  totalPrice,
 };
